fixing vectoangles(), angles now NOT rounded to integers, should make third person cam more smooth, well at least my aimbot not suffer from jerkiness anymore. Should do similar fix to ktx/mvdsv. I see it already fixed in FTE.

+++ cl_cam.c
@@ -85,12 +85,12 @@ void vectoangles(vec3_t vec, vec3_t ang) {
 		yaw = 0;
 		pitch = (vec[2] > 0) ? 90 : 270;
 	} else {
-		yaw = (int) (atan2(vec[1], vec[0]) * 180 / M_PI);
+		yaw = /*(int)*/ (atan2(vec[1], vec[0]) * 180 / M_PI);
 		if (yaw < 0)
 			yaw += 360;
 
 		forward = sqrt (vec[0]*vec[0] + vec[1]*vec[1]);
-		pitch = (int) (atan2(vec[2], forward) * 180 / M_PI);
+		pitch = /*(int)*/ (atan2(vec[2], forward) * 180 / M_PI);
 		if (pitch < 0)
 			pitch += 360;
 	}
